trunner: add the interface for adding long tests
JIRA: PD-292
